+++ aiohttp_middlewares/__init__.py
@@ -7,11 +7,16 @@ Collection of useful middlewares for aiohttp applications.
 
 """
 
-__version__ = "1.0.0"
+__version__ = "1.1.0a0"
 
 from .constants import IDEMPOTENT_METHODS, NON_IDEMPOTENT_METHODS
 from .cors import cors_middleware
-from .error import default_error_handler, error_context, error_middleware
+from .error import (
+    default_error_handler,
+    error_context,
+    error_middleware,
+    get_error_response,
+)
 from .https import https_middleware
 from .shield import shield_middleware
 from .timeout import timeout_middleware
@@ -23,6 +28,7 @@ from .utils import match_path
     default_error_handler,
     error_context,
     error_middleware,
+    get_error_response,
     https_middleware,
     IDEMPOTENT_METHODS,
     match_path,